server/DjangoBackend/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.6
django-filter==24.3
djangorestframework==3.15.2
environs==14.1.0
filelock==3.17.0
marshmallow==3.26.1
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.29.1
drf-yasg==1.21.9
psycopg2-binary